fork flot into Kibana
The flot-charts project has been abandoned since 2014, but we use it
pretty extensively in Kibana. By forking it, we can fix bugs and improve
on the library in ways that we need to for Kibana.

The goal here is not to create a replacement for flot in the general
sense but instead to use flot as the basis of a Kibana-specific charting
tool.